+++ rules/prettier/no-prettier.js
@@ -67,7 +67,10 @@ module.exports = {
       'error',
       { functions: 'always', keywords: 'always', classes: 'always' }
     ],
-    'space-before-function-paren': ['error', 'never'],
+    'space-before-function-paren': [
+      'error',
+      { anonymous: 'never', named: 'never', asyncArrow: 'always' }
+    ],
     'space-in-parens': ['warn', 'always'],
     'space-infix-ops': 'error',
     'space-unary-ops': ['error', { words: true, nonwords: false }],